it's a mess :sob: but it works :partying_face:

+++ src/global/protocol_structures/instructions.rs
@@ -691,6 +691,6 @@ pub struct TypeMetadata {
 #[derive(BinRead, BinWrite, Clone, Debug, PartialEq)]
 #[brw(little)]
 pub struct RangeData {
-    pub start: u8,
-    pub end: u8,
+    pub start: i64,
+    pub end: i64,
 }
